Fix RandomState initialization to take long integers.
The numpy.random.RandomState object does not take long integers
(more than 2**32-1) as input, although it will take as input
a numpy.array of int64's to initialize the RandomState, as long
as each element of that array is between 0 and 2**32-1, inclusive.
This can be problematic for our audit, as the audit seed may well
be a 20-digit integer.

Thus, utils.py now has the routine
    utils.RandomState(v)
which takes as input an arbitrarily large nonnegative integer v,
and returns a numpy.random.RandomState object initialized via
an a numpy.array object initialized from v.